Its about restricted entry queue
The algorithm of entering and leaving the queue (only the head of the queue is allowed to output) is implemented on the sequential storage structure. Let each element represent a job to be processed, and the element value represent the expected time of the job. If the expected execution time of a newly submitted job is less than the average time of the head and tail jobs, it will be inserted at the head of the queue, otherwise it will be inserted at the tail of the queue.